LJSUP-11064: scheduled entries posting worker has been updated.
git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@21004 1dfba125-df10-0410-8a77-af1a8d6d074f

+++ cgi-bin/LJ/DelayedEntry/Scheduler.pm
@@ -1,12 +1,53 @@
+package LJ::DelayedEntry::Scheduler::TableLock;
+use strict;
+use warnings;
+
+my $DELAYED_ENTRIES_LOCK_NAME = 'delayed_entries_lock';
+
+sub new {
+    my ($class, $dbh) = @_;
+
+    if (try_lock($dbh)) {
+        return undef;
+    }
+
+    my $self = bless {}, $class;
+    $self->{dbh} = $dbh;
+
+    return $self;
+}
+
+sub try_lock {
+    my ($dbh) = @_;
+
+    my ($free) = 
+        $dbh->selectrow_array("SELECT IS_FREE_LOCK('$DELAYED_ENTRIES_LOCK_NAME')");
+
+    if (!$free) {
+        return 0;
+    }
+
+    my ($result) = 
+        $dbh->selectrow_array("SELECT GET_LOCK('$DELAYED_ENTRIES_LOCK_NAME', 10)");
+    return $result;
+}
+
+sub DESTROY {
+    my ($self) = @_;
+    my $dbh = $self->{dbh};
+
+    $dbh->selectrow_array("SELECT RELEASE_LOCK('$DELAYED_ENTRIES_LOCK_NAME')");
+}
+
+
 package LJ::DelayedEntry::Scheduler;
 use LJ::DelayedEntry;
 use LJ::Text;
 
 use strict;
 use warnings;
-use Data::Dumper;
 
-my $PULSE_TIME = 1 * 60; 
+my $PULSE_TIME = 1;
 
 sub pulse_time {
     return $PULSE_TIME;
@@ -18,7 +59,7 @@ sub __load_delayed_entries {
     
     my $list = $dbh->selectall_arrayref("SELECT journalid, delayedid, posterid " .
                                         "FROM delayedlog2 ".
-                                        "WHERE posttime <= NOW()");
+                                        "WHERE posttime <= NOW() LIMIT 1000");
 
     foreach my $tuple (@$list) {
         push @entries, LJ::DelayedEntry->load_data($dbh,
@@ -46,23 +87,38 @@ sub __send_error {
     });    
 }
 
+
 sub on_pulse {
     my ($clusterid, $dbh) = @_;
     __assert($dbh);
-    my $entries = __load_delayed_entries($dbh);
 
-    foreach my $entry(@$entries) {
-        my $post_status = $entry->convert();
+    my $lock = new LJ::DelayedEntry::Scheduler::TableLock($dbh);
 
-        # do we need to send error
-        if ( $post_status->{error_message} ) {
-            __send_error($entry->poster, 
-                         $entry->data->{subject},
-                         $post_status->{error_message});
-        }
-        if ( $post_status->{delete_entry} ) {
-            $entry->delete();
-        }
+    if (!$lock) {
+        return;
+    }
+
+    eval {
+        my $entries = __load_delayed_entries($dbh);
+
+        do {
+            foreach my $entry(@$entries) {
+                my $post_status = $entry->convert();
+        
+                # do we need to send error
+                if ( $post_status->{error_message} ) {
+                    __send_error($entry->poster, 
+                                $entry->data->{subject},
+                                $post_status->{error_message});
+                }
+                if ( $post_status->{delete_entry} ) {
+                    $entry->delete();
+                }
+            }
+        } while ($entries = __load_delayed_entries($dbh) && @$entries)
+    };
+    if ($@) {
+        warn 'worker has failed: ' . $@;
     }
 }
